loop over locations
