--- lambdas/show-data-lambda/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "rockrobot",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.mjs",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "aws-sdk": "^2.1659.0",
-    "cloudscraper": "^4.6.0"
-  }
-}
